+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -->
 
 ## [Unreleased]
+### Added
+- Export PDF with styled free text units. [#997](https://github.com/geli-lms/geli/issues/997)
+- More predefined custom containers. [#996](https://github.com/geli-lms/geli/issues/996)
+- Styled code snippets. [#1017](https://github.com/geli-lms/geli/issues/1017)
 
 ## [[0.8.3](https://github.com/geli-lms/geli/releases/tag/v0.8.3)] - 2018-11-29 - WS 18/19 🚀-Release
 ### Added